virt-contoller: move network related code to network dir
Add OWNERS file to this dir.
