+++ pkg/virt-controller/watch/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "application.go",
         "migration.go",
         "migrationpolicy.go",
-        "network.go",
         "node.go",
         "pool.go",
         "replicaset.go",
@@ -48,6 +47,7 @@ go_library(
         "//pkg/virt-api/webhooks:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/leaderelectionconfig:go_default_library",
+        "//pkg/virt-controller/network:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/clone:go_default_library",
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
@@ -108,7 +108,6 @@ go_test(
     srcs = [
         "application_test.go",
         "migration_test.go",
-        "network_test.go",
         "node_test.go",
         "pool_test.go",
         "replicaset_test.go",
@@ -132,6 +131,7 @@ go_test(
         "//pkg/storage/types:go_default_library",
         "//pkg/testutils:go_default_library",
         "//pkg/virt-config:go_default_library",
+        "//pkg/virt-controller/network:go_default_library",
         "//pkg/virt-controller/services:go_default_library",
         "//pkg/virt-controller/watch/clone:go_default_library",
         "//pkg/virt-controller/watch/drain/disruptionbudget:go_default_library",
